refactor: decouple verifier-v2 and encoder (#2147)
* refactor: decouple Verifier-v2 and Encoder

Verifier-v2 held an encoder as subfield, just to call Decode(). There is no reason for this. Decoding has nothing to do with verification, and they have different dependencies. Someone who needs to do both should just hold on to a verifier and a decoder.

* refactor(verifier-v2): remove unnecessary kzgConfig field

* fix lint

* rename verifier -> encoder

* fix encoding/bench/main

* address litt3 comment: make TargetNumChunks uint64

+++ core/test/core_test.go
@@ -151,7 +151,7 @@ func prepareBatch(t *testing.T, operatorCount uint, blobs []core.Blob, bn uint) 
 				t.Fatal(err)
 			}
 
-			params := encoding.ParamsFromMins(chunkLength, info.TotalChunks)
+			params := encoding.ParamsFromMins(uint64(chunkLength), info.TotalChunks)
 
 			commitments, chunks, err := p.EncodeAndProve(blob.Data, params)
 			if err != nil {
